+++ crypto/src/androidTest/java/com/frybits/harmony/secure/test/EncryptedHarmonyTest.kt
@@ -350,4 +350,48 @@ class EncryptedHarmonyTest {
         assertNull(unencryptedPrefs.getStringSet("test", null)) // This key is encrypted, so it should return null
         assertEquals(randomStringSet, sharedPreferences.getStringSet("test", null))
     }
+
+    @Test
+    fun testValueToNullContains() {
+        // Everything should be empty
+        assertFalse { sharedPreferences.contains("test") }
+
+        assertEquals(0, sharedPreferences.all.size)
+
+        val randomString = "${Random.nextInt()}"
+        sharedPreferences.edit { putString("test", randomString) }
+
+        assertTrue { sharedPreferences.contains("test") }
+        assertEquals(1, sharedPreferences.all.size)
+
+        sharedPreferences.edit { putString("test", null) }
+
+        assertFalse { sharedPreferences.contains("test") }
+    }
+
+    @Test
+    fun testRemove() {
+        // Everything should be empty
+        assertFalse { sharedPreferences.contains("test") }
+
+        assertEquals(0, sharedPreferences.all.size)
+
+        val randomString = "${Random.nextInt()}"
+        sharedPreferences.edit {
+            putString("test", randomString)
+            putString(null, randomString)
+        }
+
+        assertTrue { sharedPreferences.contains("test") }
+        assertTrue { sharedPreferences.contains(null) }
+        assertEquals(2, sharedPreferences.all.size)
+
+        sharedPreferences.edit { remove("test") }
+        assertFalse { sharedPreferences.contains("test") }
+        assertTrue { sharedPreferences.contains(null) }
+
+        sharedPreferences.edit { remove(null) }
+        assertFalse { sharedPreferences.contains("test") }
+        assertFalse { sharedPreferences.contains(null) }
+    }
 }